+++ earth2studio/data/arco.py
@@ -25,15 +25,18 @@ import gcsfs
 import numpy as np
 import xarray as xr
 import zarr
+from fsspec.implementations.cached import WholeFileCacheFileSystem
 from loguru import logger
 from modulus.distributed.manager import DistributedManager
 from tqdm import tqdm
 
-from earth2studio.data.utils import prep_data_inputs, unordered_generator
+from earth2studio.data.utils import (
+    datasource_cache_root,
+    prep_data_inputs,
+    unordered_generator,
+)
 from earth2studio.lexicon import ARCOLexicon
 from earth2studio.utils.type import TimeArray, VariableArray
-
-LOCAL_CACHE = os.path.join(os.path.expanduser("~"), ".cache", "earth2studio")
 
 
 class ARCO:
@@ -68,20 +71,24 @@ class ARCO:
         self._cache = cache
         self._verbose = verbose
 
+        fs = gcsfs.GCSFileSystem(
+            cache_timeout=-1,
+            token="anon",  # noqa: S106
+            access="read_only",
+            block_size=2**20,
+        )
+
         if self._cache:
-            gcstore = fsspec.get_mapper(
-                "gs://gcp-public-data-arco-era5/ar/full_37-1h-0p25deg-chunk-1.zarr-v3",
-                target_protocol="gs",
-                cache_storage=self.cache,
-                target_options={"anon": True, "default_block_size": 2**20},
-            )
-        else:
-            gcs = gcsfs.GCSFileSystem(cache_timeout=-1)
-            gcstore = gcsfs.GCSMap(
-                "gs://gcp-public-data-arco-era5/ar/full_37-1h-0p25deg-chunk-1.zarr-v3",
-                gcs=gcs,
-            )
-        self.zarr_group = zarr.open(gcstore, mode="r")
+            cache_options = {
+                "cache_storage": self.cache,
+                "expiry_time": 31622400,  # 1 year
+            }
+            fs = WholeFileCacheFileSystem(fs=fs, **cache_options)
+
+        fs_map = fsspec.FSMap(
+            "gcp-public-data-arco-era5/ar/full_37-1h-0p25deg-chunk-1.zarr-v3", fs
+        )
+        self.zarr_group = zarr.open(fs_map, mode="r")
         self.async_timeout = 600
         self.async_process_limit = 4
 
@@ -225,7 +232,7 @@ class ARCO:
     @property
     def cache(self) -> str:
         """Get the appropriate cache location."""
-        cache_location = os.path.join(LOCAL_CACHE, "arco")
+        cache_location = os.path.join(datasource_cache_root(), "arco")
         if not self._cache:
             cache_location = os.path.join(
                 cache_location, f"tmp_{DistributedManager().rank}"