+++ earth2studio/data/cds.py
@@ -31,14 +31,12 @@ from loguru import logger
 from modulus.distributed.manager import DistributedManager
 from tqdm import tqdm
 
-from earth2studio.data.utils import prep_data_inputs
+from earth2studio.data.utils import datasource_cache_root, prep_data_inputs
 from earth2studio.lexicon import CDSLexicon
 from earth2studio.utils.type import TimeArray, VariableArray
 
 logger.remove()
 logger.add(lambda msg: tqdm.write(msg, end=""), colorize=True)
-
-LOCAL_CACHE = os.path.join(os.path.expanduser("~"), ".cache", "earth2studio")
 
 
 @dataclass
@@ -322,7 +320,7 @@ class CDS:
     @property
     def cache(self) -> str:
         """Get the appropriate cache location."""
-        cache_location = os.path.join(LOCAL_CACHE, "cds")
+        cache_location = os.path.join(datasource_cache_root(), "cds")
         if not self._cache:
             cache_location = os.path.join(
                 cache_location, f"tmp_{DistributedManager().rank}"