+++ earth2studio/data/wb2.py
@@ -25,15 +25,14 @@ import gcsfs
 import numpy as np
 import xarray as xr
 import zarr
+from fsspec.implementations.cached import WholeFileCacheFileSystem
 from loguru import logger
 from modulus.distributed.manager import DistributedManager
 from tqdm import tqdm
 
-from earth2studio.data.utils import prep_data_inputs
+from earth2studio.data.utils import datasource_cache_root, prep_data_inputs
 from earth2studio.lexicon import WB2Lexicon
 from earth2studio.utils.type import TimeArray, VariableArray
-
-LOCAL_CACHE = os.path.join(os.path.expanduser("~"), ".cache", "earth2studio")
 
 
 class _WB2Base:
@@ -51,20 +50,22 @@ class _WB2Base:
         self._cache = cache
         self._verbose = verbose
 
+        fs = gcsfs.GCSFileSystem(
+            cache_timeout=-1,
+            token="anon",  # noqa: S106
+            access="read_only",
+            block_size=2**20,
+        )
+
         if self._cache:
-            gcstore = fsspec.get_mapper(
-                f"gs://weatherbench2/datasets/era5/{wb2_zarr_store}",
-                target_protocol="gs",
-                cache_storage=self.cache,
-                target_options={"anon": True, "default_block_size": 2**20},
-            )
-        else:
-            gcs = gcsfs.GCSFileSystem(cache_timeout=-1)
-            gcstore = gcsfs.GCSMap(
-                f"gs://weatherbench2/datasets/era5/{wb2_zarr_store}",
-                gcs=gcs,
-            )
-        self.zarr_group = zarr.open(gcstore, mode="r")
+            cache_options = {
+                "cache_storage": self.cache,
+                "expiry_time": 31622400,  # 1 year
+            }
+            fs = WholeFileCacheFileSystem(fs=fs, **cache_options)
+
+        fs_map = fsspec.FSMap(f"weatherbench2/datasets/era5/{wb2_zarr_store}", fs)
+        self.zarr_group = zarr.open(fs_map, mode="r")
 
     def __call__(
         self,
@@ -190,7 +191,7 @@ class _WB2Base:
     @property
     def cache(self) -> str:
         """Get the appropriate cache location."""
-        cache_location = os.path.join(LOCAL_CACHE, "wb2era5")
+        cache_location = os.path.join(datasource_cache_root(), "wb2era5")
         if not self._cache:
             cache_location = os.path.join(
                 cache_location, f"tmp_{DistributedManager().rank}"
@@ -430,22 +431,25 @@ class WB2Climatology:
         self._cache = cache
         self._verbose = verbose
 
+        fs = gcsfs.GCSFileSystem(
+            cache_timeout=-1,
+            token="anon",  # noqa: S106
+            access="read_only",
+            block_size=2**20,
+        )
+
         if self._cache:
-            gcstore = fsspec.get_mapper(
-                "gs://weatherbench2/datasets/era5-hourly-climatology/"
-                + climatology_zarr_store,
-                target_protocol="gs",
-                cache_storage=self.cache,
-                target_options={"anon": True, "default_block_size": 2**20},
-            )
-        else:
-            gcs = gcsfs.GCSFileSystem(cache_timeout=-1)
-            gcstore = gcsfs.GCSMap(
-                "gs://weatherbench2/datasets/era5-hourly-climatology/"
-                + climatology_zarr_store,
-                gcs=gcs,
-            )
-        self.zarr_group = zarr.open(gcstore, mode="r")
+            cache_options = {
+                "cache_storage": self.cache,
+                "expiry_time": 31622400,  # 1 year
+            }
+            fs = WholeFileCacheFileSystem(fs=fs, **cache_options)
+
+        fs_map = fsspec.FSMap(
+            f"weatherbench2/datasets/era5-hourly-climatology/{climatology_zarr_store}",
+            fs,
+        )
+        self.zarr_group = zarr.open(fs_map, mode="r")
 
     def __call__(
         self,
@@ -559,7 +563,7 @@ class WB2Climatology:
     @property
     def cache(self) -> str:
         """Get the appropriate cache location."""
-        cache_location = os.path.join(LOCAL_CACHE, "wb2")
+        cache_location = os.path.join(datasource_cache_root(), "wb2")
         if not self._cache:
             cache_location = os.path.join(
                 cache_location, f"tmp_{DistributedManager().rank}"